Move GPKG connection exec sql implementation to OGR connection base class
Exposes "Execute SQL" functionality for all OGR supported datasources
