+++ src/core/providers/ogr/qgsgeopackageproviderconnection.cpp
@@ -28,7 +28,6 @@
 #include "qgsfeedback.h"
 #include "qgsogrutils.h"
 #include "qgsfielddomain.h"
-#include "qgsdbquerylog.h"
 
 #include <QTextCodec>
 #include <QRegularExpression>
@@ -136,21 +135,6 @@ void QgsGeoPackageProviderConnection::renameVectorTable( const QString &schema, 
   {
     QgsDebugMsgLevel( QStringLiteral( "Warning: error while updating the styles, perhaps there are no styles stored in this GPKG: %1" ).arg( ex.what() ), 4 );
   }
-}
-
-QgsVectorLayer *QgsGeoPackageProviderConnection::createSqlVectorLayer( const QgsAbstractDatabaseProviderConnection::SqlVectorLayerOptions &options ) const
-{
-  QgsProviderMetadata *providerMetadata { QgsProviderRegistry::instance()->providerMetadata( QStringLiteral( "ogr" ) ) };
-  Q_ASSERT( providerMetadata );
-  QMap<QString, QVariant> decoded = providerMetadata->decodeUri( uri() );
-  decoded[ QStringLiteral( "subset" ) ] = options.sql;
-  return new QgsVectorLayer( providerMetadata->encodeUri( decoded ), options.layerName.isEmpty() ? QStringLiteral( "QueryLayer" ) : options.layerName, providerKey() );
-}
-
-QgsAbstractDatabaseProviderConnection::QueryResult QgsGeoPackageProviderConnection::execSql( const QString &sql, QgsFeedback *feedback ) const
-{
-  checkCapability( Capability::ExecuteSql );
-  return executeGdalSqlPrivate( sql, feedback );
 }
 
 void QgsGeoPackageProviderConnection::vacuum( const QString &schema, const QString &name ) const
@@ -364,146 +348,6 @@ void QgsGeoPackageProviderConnection::setDefaultCapabilities()
   };
 }
 
-QgsAbstractDatabaseProviderConnection::QueryResult QgsGeoPackageProviderConnection::executeGdalSqlPrivate( const QString &sql, QgsFeedback *feedback ) const
-{
-
-  QgsDatabaseQueryLogWrapper logWrapper( sql, uri(), providerKey(), QStringLiteral( "QgsGeoPackageProviderConnection" ), QGS_QUERY_LOG_ORIGIN );
-
-  if ( feedback && feedback->isCanceled() )
-  {
-    logWrapper.setCanceled();
-    return QgsAbstractDatabaseProviderConnection::QueryResult();
-  }
-
-  QString errCause;
-  gdal::ogr_datasource_unique_ptr hDS( GDALOpenEx( uri().toUtf8().constData(), GDAL_OF_VECTOR | GDAL_OF_UPDATE, nullptr, nullptr, nullptr ) );
-  if ( hDS )
-  {
-
-    if ( feedback && feedback->isCanceled() )
-    {
-      logWrapper.setCanceled();
-      return QgsAbstractDatabaseProviderConnection::QueryResult();
-    }
-
-    std::chrono::steady_clock::time_point begin = std::chrono::steady_clock::now();
-    OGRLayerH ogrLayer( GDALDatasetExecuteSQL( hDS.get(), sql.toUtf8().constData(), nullptr, nullptr ) );
-    std::chrono::steady_clock::time_point end = std::chrono::steady_clock::now();
-
-    // Read fields
-    if ( ogrLayer )
-    {
-
-      auto iterator = std::make_shared<QgsGeoPackageProviderResultIterator>( std::move( hDS ), ogrLayer );
-      QgsAbstractDatabaseProviderConnection::QueryResult results( iterator );
-      results.setQueryExecutionTime( std::chrono::duration_cast<std::chrono::milliseconds>( end - begin ).count() );
-
-      gdal::ogr_feature_unique_ptr fet;
-
-      if ( fet.reset( OGR_L_GetNextFeature( ogrLayer ) ), fet )
-      {
-        // pk column name
-        QString pkColumnName;
-
-        QgsFields fields { QgsOgrUtils::readOgrFields( fet.get(), QTextCodec::codecForName( "UTF-8" ) ) };
-
-        // We try to guess the table name from the FROM clause
-        thread_local const QRegularExpression tableNameRegexp { QStringLiteral( R"re((?<=from|join)\s+(\w+)|"([^"]+)")re" ), QRegularExpression::PatternOption::CaseInsensitiveOption };
-        const auto match { tableNameRegexp.match( sql ) };
-        if ( match.hasMatch() )
-        {
-          pkColumnName = primaryKeyColumnName( match.captured( match.lastCapturedIndex() ) );
-        }
-
-        // fallback to "fid"
-        if ( pkColumnName.isEmpty() )
-        {
-          pkColumnName = QStringLiteral( "fid" );
-        }
-
-        // geom column name
-        QString geomColumnName;
-
-        OGRFeatureDefnH featureDef = OGR_F_GetDefnRef( fet.get() );
-
-        if ( featureDef )
-        {
-          if ( OGR_F_GetGeomFieldCount( fet.get() ) > 0 )
-          {
-            OGRGeomFieldDefnH geomFldDef { OGR_F_GetGeomFieldDefnRef( fet.get(), 0 ) };
-            if ( geomFldDef )
-            {
-              geomColumnName = OGR_GFld_GetNameRef( geomFldDef );
-            }
-          }
-        }
-
-        // May need to prepend PK and append geometry to the columns
-        if ( ! pkColumnName.isEmpty() )
-        {
-          const QRegularExpression pkRegExp { QStringLiteral( R"(^select\s+(\*|%1)[,\s+](.*)from)" ).arg( pkColumnName ),  QRegularExpression::PatternOption::CaseInsensitiveOption };
-          if ( pkRegExp.match( sql.trimmed() ).hasMatch() )
-          {
-            iterator->setPrimaryKeyColumnName( pkColumnName );
-            results.appendColumn( pkColumnName );
-          }
-        }
-
-        // Add other fields
-        for ( const auto &f : std::as_const( fields ) )
-        {
-          results.appendColumn( f.name() );
-        }
-
-        // Append geom
-        if ( ! geomColumnName.isEmpty() )
-        {
-          results.appendColumn( geomColumnName );
-          iterator->setGeometryColumnName( geomColumnName );
-        }
-
-        iterator->setFields( fields );
-      }
-
-      // Check for errors
-      if ( CE_Failure == CPLGetLastErrorType() || CE_Fatal == CPLGetLastErrorType() )
-      {
-        errCause = CPLGetLastErrorMsg( );
-      }
-
-      if ( ! errCause.isEmpty() )
-      {
-        logWrapper.setError( errCause );
-        throw QgsProviderConnectionException( QObject::tr( "Error executing SQL statement %1: %2" ).arg( sql, errCause ) );
-      }
-
-      OGR_L_ResetReading( ogrLayer );
-      iterator->nextRow();
-
-      return results;
-    }
-
-    // Check for errors
-    if ( CE_Failure == CPLGetLastErrorType() || CE_Fatal == CPLGetLastErrorType() )
-    {
-      errCause = CPLGetLastErrorMsg( );
-    }
-
-  }
-  else
-  {
-    errCause = QObject::tr( "There was an error opening GPKG %1!" ).arg( uri() );
-  }
-
-  if ( !errCause.isEmpty() )
-  {
-    logWrapper.setError( errCause );
-    throw QgsProviderConnectionException( QObject::tr( "Error executing SQL %1: %2" ).arg( sql, errCause ) );
-  }
-
-  return QgsAbstractDatabaseProviderConnection::QueryResult();
-}
-
 QString QgsGeoPackageProviderConnection::primaryKeyColumnName( const QString &table ) const
 {
   QString pkName;
@@ -543,86 +387,6 @@ QString QgsGeoPackageProviderConnection::primaryKeyColumnName( const QString &ta
   }
 
   return pkName;
-}
-
-QVariantList QgsGeoPackageProviderResultIterator::nextRowPrivate()
-{
-  const QVariantList currentRow = mNextRow;
-  mNextRow = nextRowInternal();
-  return currentRow;
-}
-
-QVariantList QgsGeoPackageProviderResultIterator::nextRowInternal()
-{
-  QVariantList row;
-  if ( mHDS && mOgrLayer )
-  {
-    gdal::ogr_feature_unique_ptr fet;
-    if ( fet.reset( OGR_L_GetNextFeature( mOgrLayer ) ), fet )
-    {
-      // PK
-      if ( ! mPrimaryKeyColumnName.isEmpty() )
-      {
-        row.push_back( OGR_F_GetFID( fet.get() ) );
-      }
-
-      if ( ! mFields.isEmpty() )
-      {
-        QgsFeature f { QgsOgrUtils::readOgrFeature( fet.get(), mFields, QTextCodec::codecForName( "UTF-8" ) ) };
-        const QgsAttributes constAttrs  = f.attributes();
-        for ( const QVariant &attribute : constAttrs )
-        {
-          row.push_back( attribute );
-        }
-
-        // Geom goes last
-        if ( ! mGeometryColumnName.isEmpty( ) )
-        {
-          row.push_back( f.geometry().asWkt() );
-        }
-
-      }
-      else // Fallback to strings
-      {
-        for ( int i = 0; i < OGR_F_GetFieldCount( fet.get() ); i++ )
-        {
-          row.push_back( QVariant( QString::fromUtf8( OGR_F_GetFieldAsString( fet.get(), i ) ) ) );
-        }
-      }
-    }
-    else
-    {
-      // Release the resources
-      GDALDatasetReleaseResultSet( mHDS.get(), mOgrLayer );
-      mHDS.release();
-    }
-  }
-  return row;
-}
-
-bool QgsGeoPackageProviderResultIterator::hasNextRowPrivate() const
-{
-  return ! mNextRow.isEmpty();
-}
-
-long long QgsGeoPackageProviderResultIterator::rowCountPrivate() const
-{
-  return  mRowCount;
-}
-
-void QgsGeoPackageProviderResultIterator::setFields( const QgsFields &fields )
-{
-  mFields = fields;
-}
-
-void QgsGeoPackageProviderResultIterator::setGeometryColumnName( const QString &geometryColumnName )
-{
-  mGeometryColumnName = geometryColumnName;
-}
-
-void QgsGeoPackageProviderResultIterator::setPrimaryKeyColumnName( const QString &primaryKeyColumnName )
-{
-  mPrimaryKeyColumnName = primaryKeyColumnName;
 }
 
 QgsFields QgsGeoPackageProviderConnection::fields( const QString &schema, const QString &table ) const
@@ -1131,40 +895,10 @@ QMultiMap<Qgis::SqlKeywordCategory, QStringList> QgsGeoPackageProviderConnection
   } );
 }
 
-QgsAbstractDatabaseProviderConnection::SqlVectorLayerOptions QgsGeoPackageProviderConnection::sqlOptions( const QString &layerSource )
+QString QgsGeoPackageProviderConnection::databaseQueryLogIdentifier() const
 {
-  SqlVectorLayerOptions options;
-  QgsProviderMetadata *providerMetadata { QgsProviderRegistry::instance()->providerMetadata( QStringLiteral( "ogr" ) ) };
-  Q_ASSERT( providerMetadata );
-  QMap<QString, QVariant> decoded = providerMetadata->decodeUri( layerSource );
-  if ( decoded.contains( QStringLiteral( "subset" ) ) )
-  {
-    options.sql = decoded[ QStringLiteral( "subset" ) ].toString();
-  }
-  else if ( decoded.contains( QStringLiteral( "layerName" ) ) )
-  {
-    options.sql = QStringLiteral( "SELECT * FROM %1" ).arg( QgsSqliteUtils::quotedIdentifier( decoded[ QStringLiteral( "layerName" ) ].toString() ) );
-  }
-  return options;
+  return QStringLiteral( "QgsGeoPackageProviderConnection" );
 }
 
-QgsGeoPackageProviderResultIterator::QgsGeoPackageProviderResultIterator( gdal::ogr_datasource_unique_ptr hDS, OGRLayerH ogrLayer )
-  : mHDS( std::move( hDS ) )
-  , mOgrLayer( ogrLayer )
-{
-  if ( mOgrLayer )
-  {
-    // Do not scan the layer!
-    mRowCount = OGR_L_GetFeatureCount( mOgrLayer, false );
-  }
-}
-
-QgsGeoPackageProviderResultIterator::~QgsGeoPackageProviderResultIterator()
-{
-  if ( mHDS )
-  {
-    GDALDatasetReleaseResultSet( mHDS.get(), mOgrLayer );
-  }
-}
 
 ///@endcond
