+++ src/core/providers/ogr/qgsgeopackageproviderconnection.h
@@ -24,36 +24,6 @@
 ///@cond PRIVATE
 #define SIP_NO_FILE
 
-
-
-struct QgsGeoPackageProviderResultIterator: public QgsAbstractDatabaseProviderConnection::QueryResult::QueryResultIterator
-{
-
-    QgsGeoPackageProviderResultIterator( gdal::ogr_datasource_unique_ptr hDS, OGRLayerH ogrLayer );
-
-    ~QgsGeoPackageProviderResultIterator();
-
-    void setFields( const QgsFields &fields );
-    void setGeometryColumnName( const QString &geometryColumnName );
-    void setPrimaryKeyColumnName( const QString &primaryKeyColumnName );
-
-  private:
-
-    gdal::ogr_datasource_unique_ptr mHDS;
-    OGRLayerH mOgrLayer;
-    QgsFields mFields;
-    QVariantList mNextRow;
-    QString mGeometryColumnName;
-    QString mPrimaryKeyColumnName;
-    long long mRowCount = -1;
-
-    QVariantList nextRowPrivate() override;
-    bool hasNextRowPrivate() const override;
-    long long rowCountPrivate() const override;
-    QVariantList nextRowInternal();
-
-};
-
 class QgsGeoPackageProviderConnection : public QgsOgrProviderConnection
 {
   public:
@@ -70,8 +40,6 @@ class QgsGeoPackageProviderConnection : public QgsOgrProviderConnection
     QString tableUri( const QString &schema, const QString &name ) const override;
     void dropRasterTable( const QString &schema, const QString &name ) const override;
     void renameVectorTable( const QString &schema, const QString &name, const QString &newName ) const override;
-    QgsVectorLayer *createSqlVectorLayer( const SqlVectorLayerOptions &options ) const override;
-    QueryResult execSql( const QString &sql, QgsFeedback *feedback = nullptr ) const override;
     void vacuum( const QString &schema, const QString &name ) const override;
     void createSpatialIndex( const QString &schema, const QString &name, const QgsAbstractDatabaseProviderConnection::SpatialIndexOptions &options = QgsAbstractDatabaseProviderConnection::SpatialIndexOptions() ) const override;
     bool spatialIndexExists( const QString &schema, const QString &name, const QString &geometryColumn ) const override;
@@ -81,15 +49,14 @@ class QgsGeoPackageProviderConnection : public QgsOgrProviderConnection
     QIcon icon() const override;
     QgsFields fields( const QString &schema, const QString &table ) const override;
     QMultiMap<Qgis::SqlKeywordCategory, QStringList> sqlDictionary() override;
-    SqlVectorLayerOptions sqlOptions( const QString &layerSource ) override;
+
+  protected:
+    QString databaseQueryLogIdentifier() const override;
+    QString primaryKeyColumnName( const QString &table ) const override;
 
   private:
 
     void setDefaultCapabilities();
-    //! Use GDAL to execute SQL
-    QueryResult executeGdalSqlPrivate( const QString &sql, QgsFeedback *feedback = nullptr ) const;
-    //! Returns PK name for table
-    QString primaryKeyColumnName( const QString &table ) const;
 
 };
