+++ src/core/providers/ogr/qgsogrproviderconnection.h
@@ -23,6 +23,37 @@
 ///@cond PRIVATE
 #define SIP_NO_FILE
 
+
+
+struct QgsOgrProviderResultIterator: public QgsAbstractDatabaseProviderConnection::QueryResult::QueryResultIterator
+{
+
+    QgsOgrProviderResultIterator( gdal::ogr_datasource_unique_ptr hDS, OGRLayerH ogrLayer );
+
+    ~QgsOgrProviderResultIterator();
+
+    void setFields( const QgsFields &fields );
+    void setGeometryColumnName( const QString &geometryColumnName );
+    void setPrimaryKeyColumnName( const QString &primaryKeyColumnName );
+
+  private:
+
+    gdal::ogr_datasource_unique_ptr mHDS;
+    OGRLayerH mOgrLayer;
+    QgsFields mFields;
+    QVariantList mNextRow;
+    QString mGeometryColumnName;
+    QString mPrimaryKeyColumnName;
+    long long mRowCount = -1;
+
+    QVariantList nextRowPrivate() override;
+    bool hasNextRowPrivate() const override;
+    long long rowCountPrivate() const override;
+    QVariantList nextRowInternal();
+
+};
+
+
 /**
  * \ingroup core
  * \class QgsOgrProviderConnection
@@ -44,6 +75,8 @@ class QgsOgrProviderConnection : public QgsAbstractDatabaseProviderConnection
     void remove( const QString &name ) const override;
     QString tableUri( const QString &schema, const QString &name ) const override;
     QgsAbstractDatabaseProviderConnection::TableProperty table( const QString &schema, const QString &table ) const override;
+    QueryResult execSql( const QString &sql, QgsFeedback *feedback = nullptr ) const override;
+    QgsVectorLayer *createSqlVectorLayer( const SqlVectorLayerOptions &options ) const override;
     void createVectorTable( const QString &schema, const QString &name, const QgsFields &fields, QgsWkbTypes::Type wkbType, const QgsCoordinateReferenceSystem &srs, bool overwrite, const QMap<QString, QVariant> *options ) const override;
     void dropVectorTable( const QString &schema, const QString &name ) const override;
     QList<QgsVectorDataProvider::NativeType> nativeTypes() const override;
@@ -51,10 +84,18 @@ class QgsOgrProviderConnection : public QgsAbstractDatabaseProviderConnection
     QgsFieldDomain *fieldDomain( const QString &name ) const override;
     void setFieldDomainName( const QString &fieldName, const QString &schema, const QString &tableName, const QString &domainName ) const override;
     void addFieldDomain( const QgsFieldDomain &domain, const QString &schema ) const override;
+    SqlVectorLayerOptions sqlOptions( const QString &layerSource ) override;
 
   protected:
 
     void setDefaultCapabilities();
+
+    virtual QString databaseQueryLogIdentifier() const;
+
+    virtual QString primaryKeyColumnName( const QString &table ) const;
+
+    //! Use GDAL to execute SQL
+    QueryResult executeGdalSqlPrivate( const QString &sql, QgsFeedback *feedback = nullptr ) const;
 
   private:
     bool mSingleTableDataset = false;
